backend/pyproject.toml
[project]
name = "renqing-backend"
version = "1.0.0"
description = "人情簿后端 API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn[standard]==0.27.1",
    "motor>=3.6.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.12.0",
    "python-multipart==0.0.9",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "bcrypt>=4.0,<5.0",
]
